Fix 5 new review comments: use R11 for address, fix stack alignment, clean up comments, fix error return, fix type signatures

+++ vm/ubpf_int.h
@@ -198,14 +198,14 @@ ubpf_dispatch_to_external_helper(
 int
 ubpf_jit_bounds_check(
     const struct ubpf_vm* vm,
-    uint64_t addr,
-    uint64_t size,
+    uintptr_t addr,
+    size_t size,
     const char* type,
-    uint64_t cur_pc,
+    uint16_t cur_pc,
     void* mem,
-    uint64_t mem_len,
+    size_t mem_len,
     void* stack,
-    uint64_t stack_len);
+    size_t stack_len);
 
 /**
  * @brief Fetch the instruction at the given index.
